Move generator main class to new module: swift-generator-cli
