+++ swift-generator/pom.xml
@@ -23,7 +23,7 @@
       <groupId>com.facebook.swift</groupId>
       <artifactId>swift-idl-parser</artifactId>
     </dependency>
-    
+
     <dependency>
       <groupId>org.antlr</groupId>
       <artifactId>stringtemplate</artifactId>
