+++ swift-generator/src/main/java/com/facebook/swift/generator/util/TemplateLoader.java
@@ -33,7 +33,7 @@ public class TemplateLoader
 {
     private static final Logger LOG = LoggerFactory.getLogger(TemplateLoader.class);
 
-    private static final String COMMON_TEMPLATES = "common.st";
+    private static final String COMMON_TEMPLATES = "java/common.st";
 
     private final StringTemplateErrorListener ERROR_LISTENER = new LoaderErrorListener();
